Rename files to increase consistency
The `tensorly.decomposition.parafac2` and
`tensorly.decomposition.robust_decomposition` modules provided implementations
of decompositions, but all other modules that does this are hidden (start
with a leading underscore). The files are renamed to resolve this inconsistency.
